requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.3
django-filter==1.0.4
django-jquery==3.1.0
get==0.0.0
idna==2.5
post==0.0.0
psycopg2==2.7.1
public==0.0.0
python-decouple==3.1
pytz==2017.2
query-string==0.0.0
request==0.0.0
requests==2.18.3
setupfiles==0.0.50
urllib3==1.22
virtualenv==15.1.0
whitenoise==3.3.0
